# Fix SELECT FOR XML/JSON behavior