--- a/test/JDBC/expected/forxml-subquery-vu-cleanup.out
+++ b/test/JDBC/expected/forxml-subquery-vu-cleanup.out
@@ -1,0 +1,22 @@
+-- Test BASE64 encoding on binary data
+drop view forxml_subquery_vu_v_path;
+GO
+
+drop view forxml_subquery_vu_v_base64;
+GO
+
+drop TABLE forxml_subquery_vu_t_binary;
+GO
+
+drop view forxml_subquery_vu_v_cte2;
+go
+drop view forxml_subquery_vu_v_cte3;
+go
+drop view forxml_subquery_vu_v_cte4;
+go
+drop view forxml_subquery_vu_v_correlated_subquery;
+go
+drop table forxml_subquery_vu_t_t1;
+go
+drop table forxml_subquery_vu_t_t2;
+go