SELECT FOR XML/JSON were both originally implemented by converting the select query into a function call that then called SPI_execute() against the original SELECT statement. While that implementation works for simple SELECT statements, it fails to execute properly against queries where the SELECT statement is a subquery which references CTEs or correlated aliases outside of the scope of the sub-SELECT. This is due to the fact that the SELECT statement is passed as a string argument to the internal function call for SPI. Thus, any external references are lost when SPI_execute() is called.

We resolve this issue by changing the implementation of SELECT FOR XML/JSON to use aggregate functions. This way, there is no need to store any query strings, and the executor will fully execute the sub-SELECT prior to passing the result rows to the aggregate function, allowing us to retain the context of any outer CTEs or correlations.

Signed-off-by: Jason Teng <jasonten@amazon.com>

Matching Engine PR: https://github.com/babelfish-for-postgresql/postgresql_modified_for_babelfish/pull/81

### Issues Resolved

BABEL-1876

### Testing

- [x] Regression testing against older tests
- [x] Test cases with correlated CTEs/subqueries

Performance testing results:
|# of rows | SQL Server | Old Implementation | New Implemenation
--- | --- | --- | ---
|10,000 | 10 ms | 180 ms | 176 ms|
|100,000 | 435 ms | 1816 ms | 1774 ms|
|1,000,000 | 8047 ms | 18568 ms | 18066 ms|

### Check List
- [x] Commits are signed per the DCO using --signoff 

By submitting this pull request, I confirm that my contribution is under the terms of the Apache 2.0 and PostgreSQL licenses, and grant any person obtaining a copy of the contribution permission to relicense all or a portion of my contribution to the PostgreSQL License solely to contribute all or a portion of my contribution to the PostgreSQL open source project.

For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/babelfish-for-postgresql/babelfish_extensions/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).